+++ src/main/java/com/digital/pos/application/service/ProcessOrderApplicationService.java
@@ -38,7 +38,7 @@ public class ProcessOrderApplicationService {
     return assign;
   }
 
-  public Long getLivePosition(Order savedOrder) {
+  public Integer getLivePosition(Order savedOrder) {
 
     return orderRepository.findLivePositionInQueue(savedOrder.getId()) + 1;
   }
